reinstate event control
some event sources may be random-access, while others may not. to make
this flexible, added an "event getter" member to the
EventViewerNamespace, which can be set to a callable that returns an
event requested by index (eventually gtid?). when not set, the buttons
don't work and the client gets an alarm.