stub out cmos rate monitoring
cmos rates come in as json objects via a zeromq socket. they are then
written to a DataStore (couchdb for now), and the CMOSRatesNamespace is
notified. that namespace then grabs updates and sends them off
(selectively) to clients.

to try it, run bin/snostream and cmos_test_sender and go to
http://localhost:8080/cmos. currently it expects a db called 'ss' on
localhost:5984.

the fake cmos rates are generated for channels 0-31.