add class to track each piece of data's flow
For example, there is a `Record` for `event_data`, which knows the
InputStream event data is coming from and the DataStore it lives in, if
applicable. Namespaces subscribe to the Record to receive changes, and
can access persistent data with Record.get(*kwargs), which is passed
through to the store.

This organizes the data to some extent and simplifies the inevitable case
where one input stream wants to put data into many data stores (e.g. orca zmq
into detector-level and crate-level storage).

Don't be fooled -- this isn't more  abstraction: if a store defines a
weird interface for `get`, the Namespace has to know about it.